spfs: add an option for mounting overlayfs layers using syscalls
Teach spfs to mount /spfs using direct overlayfs syscalls.
This is provided as an opt-in feature so that users can test this
alongside the current "mount"-based implementation. If no issues are
encountered then this can be made the default behavior in the future.

This approach avoids the command-line length limits that are present
when using the "mount" command to create the /spfs mount.

Related-to: #968
Related-to: #1164